+++ test/crdsvalidation/aiconfiguration.konghq.com/aigateway_consumergroup_test.go
@@ -26,8 +26,7 @@ func TestAIGatewayConsumerGroup(t *testing.T) {
 				APISpec: aiconfigurationv1alpha1.AIGatewayConsumerGroupAPISpec{
 					Name: "consumer1",
 				},
-				AIGatewayRef: commonv1alpha1.ObjectRef{
-					Type: commonv1alpha1.ObjectRefTypeNamespacedRef,
+				AIGatewayRef: aiconfigurationv1alpha1.AIGatewayRef{
 					NamespacedRef: &commonv1alpha1.NamespacedRef{
 						Name: "aigateway-1",
 					},